+++ src/main/java/net/frozenblock/wilderwild/config/WWWorldgenConfig.java
@@ -160,6 +160,7 @@ public final class WWWorldgenConfig {
 	public static final ConfigEntry<Boolean> TAIGA_BOULDER_DECORATION = CONFIG.entryBuilder("surfaceDecoration/taigaBoulders", EntryType.BOOL, true).requireRestart().build();
 	public static final ConfigEntry<Boolean> LAKE_DECORATION = CONFIG.entryBuilder("surfaceDecoration/lakes", EntryType.BOOL, true).requireRestart().build();
 	public static final ConfigEntry<Boolean> BASIN_DECORATION = CONFIG.entryBuilder("surfaceDecoration/basins", EntryType.BOOL, true).requireRestart().build();
+	public static final ConfigEntry<Boolean> NEW_SULFUR_POOL = CONFIG.entryBuilder("surfaceDecoration/newSulfurPool", EntryType.BOOL, true).requireRestart().build();
 	public static final ConfigEntry<Boolean> NEW_SULFUR_SPRING = CONFIG.entryBuilder("surfaceDecoration/newSulfurSpring", EntryType.BOOL, true).requireRestart().build();
 	public static final ConfigEntry<Boolean> NEW_SULFUR_SPIKE = CONFIG.entryBuilder("surfaceDecoration/newSulfurSpike", EntryType.BOOL, true).requireRestart().build();